UC8/docker/mural-citacoes/backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
Flask-Cors==4.0.0
gunicorn==21.2.0
